# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ambassador Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentAdmin
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ambassadors/{id}:
    get:
      operationId: getAmbassador
      tags: [ambassadors]
      summary: Get ambassador details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ambassador detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmbassadorDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
...
  /dashboard/{shortCode}:
    get:
      operationId: getDashboardData
      tags: [dashboard]
      summary: Get public dashboard data
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardDataResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - message
        - user
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - email
        - role
        - isActive
...
    AmbassadorDetail:
      type: object
      properties:
        id:
          type: integer
        getAmbassadorUid:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        shortCode:
          type: string
          nullable: true
        status:
          type: string
        company:
          type: string
          nullable: true
        dashboardSlug:
          type: string
        dashboardToken:
          type: string
        iframeUrl:
          type: string
        dashboardAccountCreated:
          type: boolean
        lastSyncedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        totalReferrals:
          type: integer
        professionalReferrals:
          type: integer
        companyReferrals:
          type: integer
        approvedReferrals:
          type: integer
        totalProfessionalShiftsWorked:
          type: integer
        totalCompanyShiftsFilled:
          type: integer
        leadCount:
          type: integer
        countClicks:
          type: integer
        countShares:
          type: integer
        totalMoneyEarned:
          type: string
          nullable: true
        moneyPaid:
          type: string
          nullable: true
        moneyPending:
          type: string
          nullable: true
        balanceMoney:
          type: string
          nullable: true
        enrolledAt:
          type: string
          nullable: true
      required:
        - id
        - getAmbassadorUid
        - email
        - firstName
        - lastName
        - shortCode
        - status
        - dashboardSlug
        - dashboardToken
        - iframeUrl
        - dashboardAccountCreated
        - createdAt
        - updatedAt
        - totalReferrals
        - professionalReferrals
        - companyReferrals
        - approvedReferrals
        - totalProfessionalShiftsWorked
        - totalCompanyShiftsFilled
        - leadCount
        - countClicks
        - countShares
    AmbassadorDetailResponse:
      type: object
      properties:
        ambassador:
          $ref: "#/components/schemas/AmbassadorDetail"
      required:
        - ambassador
    Referral:
      type: object
      properties:
        id:
          type: integer
        ambassadorId:
          type: integer
        sourceType:
          type: string
          enum: [PROFESSIONAL, COMPANY]
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        jobTitle:
          type: string
          nullable: true
        associatedOfficeId:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        approvedAt:
          type: string
          nullable: true
        createdAtSource:
          type: string
          nullable: true
        numberShiftsWorked:
          type: integer
          nullable: true
        totalShiftsWorked:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - ambassadorId
        - sourceType
        - createdAt
...
    SettingsResponse:
      type: object
      properties:
        getAmbassadorApiBaseUrl:
          type: string
          nullable: true
        getAmbassadorUsername:
          type: string
          nullable: true
        getAmbassadorApiTokenSet:
          type: boolean
        appBaseUrl:
          type: string
          nullable: true
        lastSyncAt:
          type: string
          nullable: true
        syncEnabled:
          type: boolean
        connectionStatus:
          type: string
          nullable: true
        envBaseUrlSet:
          type: boolean
        envUsernameSet:
          type: boolean
        envTokenSet:
          type: boolean
      required:
        - getAmbassadorApiTokenSet
        - syncEnabled
        - envBaseUrlSet
        - envUsernameSet
        - envTokenSet
    UpdateSettingsRequest:
      type: object
      properties:
        getAmbassadorApiBaseUrl:
          type: string
        getAmbassadorUsername:
          type: string
        getAmbassadorApiToken:
          type: string
        appBaseUrl:
          type: string
        syncEnabled:
          type: boolean
...
    DashboardDataResponse:
      type: object
      properties:
        ambassador:
          type: object
          properties:
            fullName:
              type: string
            shortCode:
              type: string
            status:
              type: string
            lastSyncedAt:
              type: string
              nullable: true
          required:
            - fullName
            - shortCode
            - status
        summary:
          type: object
          properties:
            totalReferrals:
              type: integer
            professionalReferrals:
              type: integer
            companyReferrals:
              type: integer
            approvedReferrals:
              type: integer
            totalProfessionalShiftsWorked:
              type: integer
            totalCompanyShiftsFilled:
              type: integer
          required:
            - totalReferrals
            - professionalReferrals
            - companyReferrals
            - approvedReferrals
            - totalProfessionalShiftsWorked
            - totalCompanyShiftsFilled
        professionalReferrals:
          type: array
          items:
            $ref: "#/components/schemas/Referral"
        companyReferrals:
          type: array
          items:
            $ref: "#/components/schemas/Referral"
      required:
        - ambassador
        - summary
        - professionalReferrals
        - companyReferrals
...