# lib/api-zod/src/generated/types/dashboardDataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ambassador Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardDataResponseAmbassador } from "./dashboardDataResponseAmbassador";
import type { DashboardDataResponseSummary } from "./dashboardDataResponseSummary";
import type { Referral } from "./referral";

export interface DashboardDataResponse {
  ambassador: DashboardDataResponseAmbassador;
  summary: DashboardDataResponseSummary;
  professionalReferrals: Referral[];
  companyReferrals: Referral[];
}